# docs/openapi.yaml
openapi: 3.0.1
info:
  title: smartAC
  description: OpenAPI Specification for SmartAC
  version: 1.0.0
servers:
  - url: http://127.0.0.1:5000
paths:
...
  /settings/temperature:
    get:
      description: Get temperature displayed by device
      responses:
        '200':
          description: Received temperature displayed by device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusNumberValueResponse'
    post:
      description: Set temperature to be displayed by device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemperatureValueRequest'
      responses:
        '200':
          description: Added and received temperature to be displayed by device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusNumberValueResponse'
        '400':
          description: Missing temperature value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /settings/mode:
    get:
      description: Get mode used by device
      responses:
        '200':
          description: Received mode used by device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusValueResponse'
    post:
      description: Set mode to be used by device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModeValueRequest'
      responses:
        '200':
          description: Added and received mode to be used by device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusValueResponse'
        '400':
          description: Missing or wrong mode value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /settings/fanSpeed:
    get:
      description: Get fan speed used by device
      responses:
        '200':
          description: Received fan speed used by device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusValueResponse'
    post:
      description: Set fan speed to be used by device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FanSpeedRequest'
      responses:
        '200':
          description: Added and received fan speed to be used by device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusValueResponse'
        '400':
          description: Missing or wrong fan speed value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
...
  /settings/power:
    get:
      description: Check if device is on or off
      responses:
        '200':
          description: Received power value of device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusValueResponse'
    post:
      description: Turn the device on or off
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PowerRequest'
      responses:
        '200':
          description: The device has been turned on or off
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Invalid value or the device is in maintenance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /settings/light:
    get:
      description: Check if light is ON or OFF for device
      responses:
        '200':
          description: Received light value for device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LightResponse'
    post:
      description: Turn the light on or off
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LightRequest'
      responses:
        '200':
          description: The light is now on or off
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Wrong light value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /settings/sound:
    get:
      description: Check if sound is ON or OFF for device
      responses:
        '200':
          description: Received sound value for device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoundResponse'
    post:
      description: Turn the sound on or off
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SoundRequest'
      responses:
        '200':
          description: The sound is now on or off
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Wrong sound value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /settings/cleaning:
    get:
      description: Check if the device is in cleaning or not
      responses:
        '200':
          description: Received cleaning state of device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusValueResponse'
    post:
      description: Turn the sound on or off
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CleaningRequest'
      responses:
        '200':
          description: The cleaning state is changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Wrong or missing cleaning value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
components:
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
    StatusValueResponse:
      type: object
      properties:
        status:
          type: string
        value:
          type: string
    StatusNumberValueResponse:
      type: object
      properties:
        status:
          type: string
        value:
          type: number
    SoundResponse:
      type: object
      properties:
        status:
          type: string
        value:
          type: string
        volume:
          type: integer
    LightResponse:
      type: object
      properties:
        status:
          type: string
        value:
          type: string
        intensity:
          type: integer
    ModeValueRequest:
      type: object
      properties:
        value:
          type: string
          enum:
          - AUTO
          - COOL
          - DRY
          - FAN
          - HEAT
          - ECO
    CleaningRequest:
      type: object
      properties:
        value:
          type: string
          enum:
          - START
          - STOP
    FanSpeedRequest:
      type: object
      properties:
        value:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
    PowerRequest:
      type: object
      properties:
        value:
          type: string
          enum:
          - ON
          - OFF
    TemperatureValueRequest:
      type: object
      properties:
        value:
          type: integer
          minimum: 16
          maximum: 28
...
    LightRequest:
      type: object
      properties:
        value:
          type: string
          enum:
          - ON
          - OFF
        intensity:
          type: integer
    SoundRequest:
      type: object
      properties:
        value:
          type: string
          enum:
          - ON
          - OFF
        volume:
          type: integer
...